requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
pillow==10.1.0
opencv-python==4.8.1.78
pydub==0.25.1
speechrecognition==3.10.0
moviepy==1.0.3
whisper==20231117
librosa==0.10.1
pytube==15.0.0
py7zr==0.20.5
rarfile==4.0
unidecode==1.3.8
chardet==5.2.0